refactor: extract native request util
