+++ src/tools/get-page.ts
@@ -1,7 +1,5 @@
 import { z } from 'zod';
-import https from 'https';
-import http from 'http';
-import { URL } from 'url';
+import { makeNativeHttpRequest } from './native-request.js';
 import { GrowiClient } from '../growi-client.js';
 
 const logToStderr = (...args: any[]) => {
@@ -14,70 +12,6 @@ export const getPageSchema = z.object({
 
 export type GetPageParams = z.infer<typeof getPageSchema>;
 
-function makeNativeHttpRequest(url: string, apiToken: string): Promise<any> {
-  return new Promise((resolve, reject) => {
-    const parsedUrl = new URL(url);
-    logToStderr(`Making native HTTP request to: ${url}`);
-
-    const postData = `access_token=${encodeURIComponent(apiToken)}`;
-
-    const options = {
-      hostname: parsedUrl.hostname,
-      port: parsedUrl.port || (parsedUrl.protocol === 'https:' ? 443 : 80),
-      path: `${parsedUrl.pathname}${parsedUrl.search}`,
-      method: 'GET',
-      headers: {
-        'User-Agent': 'curl/8.7.1',
-        'Accept': '*/*',
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'Content-Length': Buffer.byteLength(postData)
-      }
-    };
-
-    logToStderr(`Sending access_token in request body: ${apiToken.substring(0, 5)}...`);
-
-    const protocol = parsedUrl.protocol === 'https:' ? https : http;
-    const req = protocol.request(options, (res) => {
-      logToStderr(`Response status: ${res.statusCode}`);
-
-      let data = '';
-      res.on('data', (chunk) => {
-        data += chunk.toString();
-      });
-
-      res.on('end', () => {
-        logToStderr(`Response completed. Data length: ${data.length}`);
-
-        if (res.statusCode && res.statusCode >= 200 && res.statusCode < 300) {
-          try {
-            const jsonData = JSON.parse(data);
-            if (jsonData.page && jsonData.page.revision) {
-              const bodyLength = jsonData.page.revision.body ? jsonData.page.revision.body.length : 0;
-              logToStderr(`Got page with content length: ${bodyLength} characters`);
-              if (bodyLength > 0) {
-                const preview = jsonData.page.revision.body.substring(0, 100);
-                logToStderr(`Content preview: ${preview}...`);
-              }
-            }
-            resolve(jsonData);
-          } catch (error) {
-            reject(new Error(`Failed to parse JSON: ${error instanceof Error ? error.message : String(error)}`));
-          }
-        } else {
-          reject(new Error(`HTTP Error: ${res.statusCode} ${res.statusMessage || ''} - ${data}`));
-        }
-      });
-    });
-
-    req.on('error', (error) => {
-      logToStderr(`Request failed: ${error.message}`);
-      reject(error);
-    });
-
-    req.write(postData);
-    req.end();
-  });
-}
 
 export async function getPage(
   client: GrowiClient,
