+++ src/tools/recently-updated-pages.ts
@@ -1,9 +1,7 @@
 import { z } from 'zod';
 import { GrowiClient } from '../growi-client.js';
 import { GrowiPage } from '../types/growi.js';
-import https from 'https';
-import http from 'http';
-import { URL } from 'url';
+import { makeNativeHttpRequest } from './native-request.js';
 
 const logToStderr = (...args: any[]) => {
   console.error(...args);
@@ -16,63 +14,6 @@ export const recentlyUpdatedPagesSchema = z.object({
 
 export type RecentlyUpdatedPagesParams = z.infer<typeof recentlyUpdatedPagesSchema>;
 
-function makeNativeHttpRequest(url: string, apiToken: string): Promise<any> {
-  return new Promise((resolve, reject) => {
-    const parsedUrl = new URL(url);
-    logToStderr(`Making native HTTP request to: ${url}`);
-
-    const postData = `access_token=${encodeURIComponent(apiToken)}`;
-
-    const options = {
-      hostname: parsedUrl.hostname,
-      port: parsedUrl.port || (parsedUrl.protocol === 'https:' ? 443 : 80),
-      path: `${parsedUrl.pathname}${parsedUrl.search}`,
-      method: 'GET',
-      headers: {
-        'User-Agent': 'curl/8.7.1',
-        'Accept': '*/*',
-        'Content-Type': 'application/x-www-form-urlencoded',
-        'Content-Length': Buffer.byteLength(postData)
-      }
-    };
-
-    logToStderr(`Sending access_token in request body: ${apiToken.substring(0, 5)}...`);
-
-    const protocol = parsedUrl.protocol === 'https:' ? https : http;
-    const req = protocol.request(options, (res) => {
-      logToStderr(`Response status: ${res.statusCode}`);
-
-      let data = '';
-      res.on('data', (chunk) => {
-        data += chunk.toString();
-      });
-
-      res.on('end', () => {
-        logToStderr(`Response completed. Data length: ${data.length}`);
-
-        if (res.statusCode && res.statusCode >= 200 && res.statusCode < 300) {
-          try {
-            const jsonData = JSON.parse(data);
-            logToStderr(`Got ${jsonData.pages?.length || 0} pages out of ${jsonData.totalCount || 0} total`);
-            resolve(jsonData);
-          } catch (error) {
-            reject(new Error(`Failed to parse JSON: ${error instanceof Error ? error.message : String(error)}`));
-          }
-        } else {
-          reject(new Error(`HTTP Error: ${res.statusCode} ${res.statusMessage || ''} - ${data}`));
-        }
-      });
-    });
-
-    req.on('error', (error) => {
-      logToStderr(`Request failed: ${error.message}`);
-      reject(error);
-    });
-
-    req.write(postData);
-    req.end();
-  });
-}
 
 export async function recentlyUpdatedPages(
   client: GrowiClient,
